requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
mss>=9.0.0
pyautogui>=0.9.54
Pillow>=10.0.0
scikit-image>=0.22.0
pywinauto>=0.6.8
psutil>=5.9.0
pyperclip>=1.8.2
aiofiles>=23.0.0
bcrypt>=4.1.0
anthropic>=0.25.0
openai>=1.30.0
google-genai>=0.5.0
PyYAML>=6.0.1
httpx>=0.27.0
html2text>=2024.2.26
